pyproject.toml
[tool.poetry]
name = "rag-books"
version = "0.1.0"
description = ""
authors = ["Artur Dernst <artur.dernst@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
sentence-transformers = "^5.0.0"
faiss-cpu = "^1.11.0"
openai = "^1.95.1"


[[tool.poetry.packages]]
include = "src"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
